requirements.txt
fastapi
uvicorn
opencv-python
onnxruntime
numpy
python-multipart
gtts
python-dotenv